Applied project renaming on parts where it was not updated by refactoring
In the commit with ID: 4416f5c607e3daef2acb7656ef922521c153503b
I tried to rename the difault "example" package to "tosp" by refactoring. But that left the naming intact in some parts of the code including in build.gradle. In this commit I'm changing "example" to "tosp" in such files.

NOTE: there are files which are not being committed but this renaming should be applied in their code as well. Find them in your project by pressing Ctrl+Shift+F in Android Studio IDE and search for "com.example.sharingexpenses" and change it to "com.tosp.sharingexpenses"